Exported Structs

Main.NonsmoothFwdAD.GeneralizedDiff.AFloat — Type

AFloat{val::Float64, dot::Vector{Float64}, ztol::Float64}

Type representing generalized derivative information.

Fields

val::Float64: output value

dot::Vector64: output generalized derivative. Set to I.

ztol::Float64: capture radius for nonsmooth operations with kinks. Set to 0.0 by default.

Notes:

Operator overloading macros defining generalized differentiation rules for elemental operations require AFloat inputs. Structs cannot be defined outside main module file.

Exported Functions

Functions from GeneralizedDiff.jl

Main.NonsmoothFwdAD.GeneralizedDiff.eval_ld_derivative — Function

eval_ld_derivative(f::Function, x::Vector{Float64}, xDot::T; ztol::Float64)

Compute the function vector-valued f(x) and the LD-derivative f'(x; xDot).

Arguments

f::Function: must be continous and of finite compositions of elemental operations. Each operation must be of the form f(x::N)::Float64 where N is either Vector{Float64} or Float64 ; otherwise implementation cannot map f. Supported elemental operations include: +, -, *, inv, /, ^, exp, log, sin, cos, abs, min, max, hypot.

x::Vector{Float64}: input domain vector

xDot::T: output generalized derivative. It must have full row rank. Set to I by default.

where T is either Matrix{Float64} or Vector{Vector{Float64}}

Keywords

ztol::Float64: capture radius for nonsmooth operations with kinks. Set to 0.0 by default.

Returns

y::Vector{Float64}:

yDot::T where T is either Matrix{Float64} or Vector{Vector{Float64}}: the lexicographic derivative f'(x; xDot)

Example

To evaluate the LD-derivative of function f on vector x:

f(x) = min(x[2]*x[3] + x[1], -x[3])
x = [0.0, 0.0, 0.0]
xDot = Matrix{Float64}(I(length(x0)))
_, yDot = eval_ld_derivative(f, x, xDot)

# output 

yDot = [0.0, 0.0, -1.0] 

Main.NonsmoothFwdAD.GeneralizedDiff.eval_dir_derivative — Function

eval_dir_derivative (f::Function, x::Vector{Float64}, xDot::Vector{Float64}; ztol::Float64)

Compute the function vector-valued y = f(x) and the directional derivative yDeriv = f'(x; xDot).

See eval_ld_derivative for more details on function inputs.

Returns

y::Vector{Float64}: function value as a vector

yDot::Vector{Float64}: directional derivative f'(x; xDot::Vector) as a vector

Notes

For this function, argument xDot must be of type Vector{Float64}.

Example

To calculate the directional derivative of function f in the direction of vector x.

f(x) = min(x[2]*x[3] + x[1], -x[3])
x = [0.0, 0.0, 0.0]
xDot = [1.0, 0.0, 1.0]
_, yDeriv = eval_dir_derivative(f, x, xDot)

# output 

yDeriv = [-1.0] 

Main.NonsmoothFwdAD.GeneralizedDiff.eval_gen_derivative — Function

eval_gen_derivative (f::Function, x::Vector{Float64}, xDot::Matrix{Float64}; ztol::Float64)

Compute the function vector-valued y = f(x) and the generalized derivative yDeriv = Df(x; xDot).

See eval_ld_derivative for more details on function inputs.

Returns

y::Vector{Float64}: function value as a vector

yDeriv::Matrix{Float64}: lexicographic derivative Df(x; xDot::Vector) as a vector

Notes

For this function, argument xDot must be of type Matrix{Float64}.

Example

To calculate the generalized derivative of function f in the direction of vector x.

f(x) = min(x[2]*x[3] + x[1], -x[3])
x = [0.0, 0.0, 0.0]
xDot = Matrix{Float64}(I(length(x0)))
_, yDeriv = eval_gen_derivative(f, x, xDot)
    
# output 
    
yDeriv = [0.0, 0.0, -1.0]] 

Main.NonsmoothFwdAD.GeneralizedDiff.eval_gen_gradient — Function

eval_gen_gradient (f::Function, x::Vector{Float64}, xDot::Matrix{Float64}; ztol::Float64)

Compute the function vector-valued y = f(x) and the generalized gradient yGrad = yDeriv' for scalar valued functions f.

See eval_ld_derivative for more details on function inputs.

Returns

y::Vector{Float64}: function value as a vector

yGrad:Vector{Float64}: directional derivative f'(x; xDot::Vector) as a vector

Notes

For this function, argument xDot must be of type Matrix{Float64}.

Example

To calculate the generalized gradient of function f in the direction of vector x.

f(x) = min(x[2]*x[3] + x[1], -x[3])
x = [0.0, 0.0, 0.0]
xDot = Matrix{Float64}(I(length(x0)))
_, yDeriv = eval_gen_gradient(f, x, xDot)
    
# output 
    
yDeriv = [0.0, 0.0, -1.0]] 

Main.NonsmoothFwdAD.GeneralizedDiff.eval_compass_difference — Function

eval_compass_difference (f::Function, x::Vector{Float64}, ztol::Float64)

Compute the function vector-valued y = f(x) and the compass difference of a scalar-valued f at x. If f has a domain dimension of 1 or 2, the compass difference is an element of Clarke's generalized gradient.

Arguments

f::Function: must be continous and of finite compositions of elemental operations. Each operation must be of the form f(x::N)::Float64 where N is either Vector{Float64} or Float64 ; otherwise implementation cannot map f. Supported elemental operations include: +, -, *, inv, /, ^, exp, log, sin, cos, abs, min, max, hypot.

x::Vector{Float64}: input domain vector

Keywords

ztol::Float64: capture radius for nonsmooth operations with kinks. Set to 0.0 by default.

Returns

y::Vector{Float64}: function value as a vector

yCompass::Vector{Float64}: function compass difference

Example

To calculate the compass difference of function f at x.

f(x) = min(x[2]*x[3] + x[1], -x[3])
x = [0.0, 0.0, 0.0]
_, yCompass = eval_compass_difference(f, x)
    
# output 
    
yCompass = [0.5, 0.0, -0.5]

Functions from ConvexOptimization.jl

Main.NonsmoothFwdAD.ConvexOptimization.semiSmoothNewton — Function

semiSmoothNewton(f::Function, x0::Vector{Float64}, kwargs…)

Compute vector x where f(x) = 0 using the semi-smooth Newton method where there are multiple inputs to f but only one output.

Arguments

f::Function: must be continous and of finite compositions of elemental operations. Each operation must be of the form f(x::N)::Float64 where N is either Vector{Float64} or Float64 ; otherwise implementation cannot map f. Supported elemental operations include: +, -, *, inv, /, ^, exp, log, sin, cos, abs, min, max, hypot.

x0::Vector{Float64}: initial domain vector guess

Keywords

epsilon::Float64: tolerance for solver stopping condition where epsilon + delta*norm(f(x0)). Set to 0.05 by default.

delta::Float64: tolerance for solver stopping condition where epsilon + delta*norm(f(x0)). Set to 0.00005 by default.

maxIter::Int64: maximum number of solver iterations. Set to 1000 by default.

Main.NonsmoothFwdAD.ConvexOptimization.LPNewton — Function

LPNewton(f::Function, z0::Vector{Float64}, kwargs…)

Compute vector z where f(z) = 0 using the Linear Program Newton method where the number of inputs to f is equivalent to number of outputs.

Arguments

f::Function: must be continous and of finite compositions of elemental operations. Each operation must be of the form f(x::N)::Float64 where N is either Vector{Float64} or Float64 ; otherwise implementation cannot map f. Supported elemental operations include: +, -, *, inv, /, ^, exp, log, sin, cos, abs, min, max, hypot.

x0::Vector{Float64}: initial domain vector guess

Keywords

Constant and tolerance keyword

epsilon::Float64: tolerance for solver stopping condition where f(z) >= epsilon. Set to 1e-2 by default.

TOLERANCE::Float64: tolerance for JuMP solver. Set to 1e-6 by default.

maxIter::Int64: maximum number of solver iterations. Set to 20 by default.

Constraint keywords

lb::Vector{Float64}: lower bound on domain vector x. Set to -Inf vector of size n by default.

ub::Vector{Float64}: upper bound on domain vector x. Set to -Inf vector of size n by default.

A::Matrix{Float64}: matrix A in inequality constraint set A*z <= b. Set to zeros vector of size n by default.

b::Vector{Float64}: matrix b in inequality constraint set A*z <= b. Set to zeros matrix of size 1x1.

Aeq::Matrix{Float64}: matrix A in equality constraint set A*z == b. Set to zeros vector of size n by default.

beq::Vector{Float64}: matrix b in equality constraint set A*z == b. Set to zeros matrix of size 1x1.

where n is the length of vector z0.

Main.NonsmoothFwdAD.ConvexOptimization.levelMethod — Function

levelMethod(f::Function, z0::Vector{Float64}, kwargs…)

Compute vector x to minimize f(x) using the Level method where there are multiple inputs to f but only one output. .

Arguments

f::Function: must be continous and of finite compositions of elemental operations. Each operation must be of the form f(x::N)::Float64 where N is either Vector{Float64} or Float64 ; otherwise implementation cannot map f. Supported elemental operations include: +, -, *, inv, /, ^, exp, log, sin, cos, abs, min, max, hypot.

x0::Vector{Float64}: initial domain vector guess

Keywords

Constant and tolerance keywords

epsilon::Float64: tolerance for solver stopping condition where f(z) >= epsilon. Set to 1e-2 by default.

TOLERANCE::Float64: tolerance for JuMP solver. Set to 1e-6 by default.

alpha::Float64: level set constant for quadratic programming. Set to 1/(2+sqrt(2) by default.

maxIter::Int64: maximum number of solver iterations. Set to 20 by default.

Constraint keywords

lb::Vector{Float64}: lower bound on domain vector x. Set to -Inf vector of size n by default.

ub::Vector{Float64}: upper bound on domain vector x. Set to -Inf vector of size n by default.

A::Matrix{Float64}: matrix A in inequality constraint set A*z <= b. Set to zeros vector of size n by default.

b::Vector{Float64}: matrix b in inequality constraint set A*z <= b. Set to zeros matrix of size 1x1.

Aeq::Matrix{Float64}: matrix A in equality constraint set A*z == b. Set to zeros vector of size n by default.

beq::Vector{Float64}: matrix b in equality constraint set A*z == b. Set to zeros matrix of size 1x1.

where n is the length of vector z0.