Exported Functions

Main.NonsmoothFwdAD.ConvexOptimization.semiSmoothNewton — Function

semiSmoothNewton(f::Function, x0::Vector{Float64}, kwargs…)

Compute vector x where f(x) = 0 using the semi-smooth Newton method where there are multiple inputs to f but only one output.

Arguments

f::Function: must be continous and of finite compositions of elemental operations. Each operation must be of the form f(x::N)::Float64 where N is either Vector{Float64} or Float64 ; otherwise implementation cannot map f. Supported elemental operations include: +, -, *, inv, /, ^, exp, log, sin, cos, abs, min, max, hypot.

x0::Vector{Float64}: initial domain vector guess

Keywords

epsilon::Float64: tolerance for solver stopping condition where epsilon + delta*norm(f(x0)). Set to 0.05 by default.

delta::Float64: tolerance for solver stopping condition where epsilon + delta*norm(f(x0)). Set to 0.00005 by default.

maxIter::Int64: maximum number of solver iterations. Set to 1000 by default.

Main.NonsmoothFwdAD.ConvexOptimization.LPNewton — Function

LPNewton(f::Function, z0::Vector{Float64}, kwargs…)

Compute vector z where f(z) = 0 using the Linear Program Newton method where the number of inputs to f is equivalent to number of outputs.

Arguments

f::Function: must be continous and of finite compositions of elemental operations. Each operation must be of the form f(x::N)::Float64 where N is either Vector{Float64} or Float64 ; otherwise implementation cannot map f. Supported elemental operations include: +, -, *, inv, /, ^, exp, log, sin, cos, abs, min, max, hypot.

x0::Vector{Float64}: initial domain vector guess

Keywords

Constant and tolerance keyword

epsilon::Float64: tolerance for solver stopping condition where f(z) >= epsilon. Set to 1e-2 by default.

solverTolerance::Float64: tolerance for JuMP solver. Set to 1e-6 by default.

maxIter::Int64: maximum number of solver iterations. Set to 20 by default.

Constraint keywords

lb::Vector{Float64}: lower bound on domain vector x. Set to -Inf vector of size n by default.

ub::Vector{Float64}: upper bound on domain vector x. Set to -Inf vector of size n by default.

A::Matrix{Float64}: matrix A in inequality constraint set A*z <= b. Set to zeros vector of size n by default.

b::Vector{Float64}: matrix b in inequality constraint set A*z <= b. Set to zeros matrix of size 1x1.

Aeq::Matrix{Float64}: matrix A in equality constraint set A*z == b. Set to zeros vector of size n by default.

beq::Vector{Float64}: matrix b in equality constraint set A*z == b. Set to zeros matrix of size 1x1.

where n is the length of vector z0.

Main.NonsmoothFwdAD.ConvexOptimization.levelMethod — Function

levelMethod(f::Function, z0::Vector{Float64}, kwargs…)

Compute vector x to minimize f(x) using the Level method where there are multiple inputs to f but only one output. .

Arguments

f::Function: must be continous and of finite compositions of elemental operations. Each operation must be of the form f(x::N)::Float64 where N is either Vector{Float64} or Float64 ; otherwise implementation cannot map f. Supported elemental operations include: +, -, *, inv, /, ^, exp, log, sin, cos, abs, min, max, hypot.

x0::Vector{Float64}: initial domain vector guess

Keywords

Constant and tolerance keywords

epsilon::Float64: tolerance for solver stopping condition where f(z) >= epsilon. Set to 1e-2 by default.

solverTolerance::Float64: tolerance for JuMP solver. Set to 1e-6 by default.

alpha::Float64: level set constant for quadratic programming. Set to 1/(2+sqrt(2) by default.

maxIter::Int64: maximum number of solver iterations. Set to 20 by default.

Constraint keywords

lb::Vector{Float64}: lower bound on domain vector x. Set to -Inf vector of size n by default.

ub::Vector{Float64}: upper bound on domain vector x. Set to -Inf vector of size n by default.

A::Matrix{Float64}: matrix A in inequality constraint set A*z <= b. Set to zeros vector of size n by default.

b::Vector{Float64}: matrix b in inequality constraint set A*z <= b. Set to zeros matrix of size 1x1.

Aeq::Matrix{Float64}: matrix A in equality constraint set A*z == b. Set to zeros vector of size n by default.

beq::Vector{Float64}: matrix b in equality constraint set A*z == b. Set to zeros matrix of size 1x1.

where n is the length of vector z0.